move find_elaz to coordinates
