+++ tests/utils/test_coordinates.py
@@ -8,8 +8,9 @@ __date__ = "21 Jun 2022"
 import numpy as np
 import pytest
 
+import gnss_lib_py.utils.constants as consts
+from gnss_lib_py.utils.coordinates import ecef_to_el_az
 from gnss_lib_py.utils.coordinates import geodetic_to_ecef, ecef_to_geodetic, LocalCoord
-
 
 @pytest.fixture(name="local_ecef")
 def fixture_local_ecef():
@@ -215,3 +216,71 @@ def test_geodetic_to_ned(local_frame):
     lla_back_t = local_frame.ned_to_geodetic(ned.T)
     np.testing.assert_array_almost_equal(lla_back, lla, decimal=0)
     np.testing.assert_array_almost_equal(lla_back, lla_back_t.T)
+
+@pytest.fixture(name="expected_elaz")
+def fixture_expected_elaz():
+    """Set the expected elevation and azimuth from sample positions.
+
+    Returns
+    -------
+    expect_elaz : np.ndarray
+        Array containing 6 el/az pairs for testing elaz function
+
+    """
+    expect_elaz = np.array([[90.0, -90.0, 0.0 ,  0.0 , 0.0, 0.0  ],
+                            [0.0 ,  0.0 , 90.0, -90.0, 0.0, 180.0]]).T
+    return expect_elaz
+
+
+@pytest.fixture(name="set_sv_pos")
+def fixture_set_sv_pos():
+    """Set the sample satellite positions for computing elevation and azimuth.
+
+    Returns
+    -------
+    sv_pos : np.ndarray
+        Array containing 6 satellite x, y, z coordinates
+
+    """
+    sv_pos = np.zeros([6, 3])
+    sv_pos[0,0] =  consts.A*1.25
+    sv_pos[1,0] =  consts.A*0.75
+    sv_pos[2,0] =  consts.A
+    sv_pos[2,1] =  consts.A
+    sv_pos[3,0] =  consts.A
+    sv_pos[3,1] = -consts.A
+    sv_pos[4,0] =  consts.A
+    sv_pos[4,2] =  consts.A
+    sv_pos[5,0] =  consts.A
+    sv_pos[5,2] = -consts.A
+    return sv_pos
+
+@pytest.fixture(name="set_rx_pos")
+def fixture_set_rx_pos():
+    """Set the sample reciever position for computing elaz.
+
+    Returns
+    -------
+    rx_pos : np.ndarray
+        Array containing 6 satellite x, y, z coordinates
+
+    """
+    rx_pos = np.reshape(np.array([consts.A, 0, 0]), [1, 3])
+    return rx_pos
+
+def test_ecef_to_el_az(expected_elaz, set_sv_pos, set_rx_pos):
+    """Test receiver to satellite azimuth and elevation calculation.
+
+    Parameters
+    ----------
+    expected_elaz : fixture
+        Expected elevation and azimuth angles generated by the given satellite
+        and receiver positions.
+    set_sv_pos : fixture
+        Satellite position setter
+    set_rx_pos : fixture
+        Receiver position setter
+
+    """
+    calc_elaz = ecef_to_el_az(set_rx_pos, set_sv_pos)
+    np.testing.assert_array_almost_equal(expected_elaz, calc_elaz)
