+++ tests/utils/test_sim_gnss.py
@@ -19,8 +19,6 @@ from gnss_lib_py.utils.coordinates import geodetic_to_ecef
 from gnss_lib_py.parsers.ephemeris import EphemerisManager
 from gnss_lib_py.utils.time_conversions import datetime_to_tow
 
-
-
 T = 0.1
 
 def timestamp():
@@ -136,56 +134,6 @@ def fixture_set_xyz():
     times = np.array([[1.], [1.], [1.]])
     return pos_array, vel_array, times
 
-@pytest.fixture(name="expected_elaz")
-def fixture_expected_elaz():
-    """Set the expected elevation and azimuth from sample positions.
-
-    Returns
-    -------
-    expect_elaz : np.ndarray
-        Array containing 6 el/az pairs for testing elaz function
-
-    """
-    expect_elaz = np.array([[90.0, -90.0, 0.0 ,  0.0 , 0.0, 0.0  ],
-                            [0.0 ,  0.0 , 90.0, -90.0, 0.0, 180.0]]).T
-    return expect_elaz
-
-@pytest.fixture(name="set_sv_pos")
-def fixture_set_sv_pos():
-    """Set the sample satellite positions for computing elevation and azimuth.
-
-    Returns
-    -------
-    sv_pos : np.ndarray
-        Array containing 6 satellite x, y, z coordinates
-
-    """
-    sv_pos = np.zeros([6, 3])
-    sv_pos[0,0] =  consts.A*1.25
-    sv_pos[1,0] =  consts.A*0.75
-    sv_pos[2,0] =  consts.A
-    sv_pos[2,1] =  consts.A
-    sv_pos[3,0] =  consts.A
-    sv_pos[3,1] = -consts.A
-    sv_pos[4,0] =  consts.A
-    sv_pos[4,2] =  consts.A
-    sv_pos[5,0] =  consts.A
-    sv_pos[5,2] = -consts.A
-    return sv_pos
-
-@pytest.fixture(name="set_rx_pos")
-def fixture_set_rx_pos():
-    """Set the sample reciever position for computing elaz.
-
-    Returns
-    -------
-    rx_pos : np.ndarray
-        Array containing 6 satellite x, y, z coordinates
-
-    """
-    rx_pos = np.reshape(np.array([consts.A, 0, 0]), [1, 3])
-    return rx_pos
-
 # Define tests
 def test_extract_xyz(set_xyz):
     """Test the extraction of satellite epoch, position, and velocity from
@@ -210,23 +158,6 @@ def test_extract_xyz(set_xyz):
     np.testing.assert_array_equal(test_pos, pos_array)
     np.testing.assert_array_equal(test_vel, vel_array)
     np.testing.assert_array_equal(prns, np.array([1, 2, 3]))
-
-def test_find_elaz(expected_elaz, set_sv_pos, set_rx_pos):
-    """Test receiver to satellite azimuth and elevation calculation.
-
-    Parameters
-    ----------
-    expected_elaz : fixture
-        Expected elevation and azimuth angles generated by the given satellite
-        and receiver positions.
-    set_sv_pos : fixture
-        Satellite position setter
-    set_rx_pos : fixture
-        Receiver position setter
-
-    """
-    calc_elaz = sim_gnss.find_elaz(set_rx_pos, set_sv_pos)
-    np.testing.assert_array_almost_equal(expected_elaz, calc_elaz)
 
 def test_measures_value_range(get_meas):
     """Test the order of magnitude of simulated pseudorange measurements,
